feat: delete screening

+++ src/app/utils/generic-crud-table/generic-crud-table.component.ts
@@ -14,6 +14,7 @@ import { MatDialogModule } from '@angular/material/dialog';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { Fields } from '../dynamic-modal-form/dynamic-modal-form';
 import { DynamicModalFormComponent } from '../dynamic-modal-form/dynamic-modal-form.component';
+import { UtilsService } from '../utils.service';
 
 @Component({
   selector: 'app-generic-crud-table',
@@ -54,7 +55,10 @@ export class GenericCrudTableComponent implements OnInit {
   currentPage: number = 1;
   totalPages: number = 1;
 
-  constructor(private translate: TranslateService) {
+  constructor(
+    private translate: TranslateService,
+    private utilsService: UtilsService
+  ) {
     this.translate.setDefaultLang('fr');
     registerLocaleData(localeFr);
   }
@@ -133,9 +137,10 @@ export class GenericCrudTableComponent implements OnInit {
 
   onModalClose(result?: any) {
     this.showModal = false;
+    console.log('result : ', result);
 
     if (result) {
-      if (result.id === 0) {
+      if (result.id === undefined || result.id === '') {
         this.addItem.emit(result);
       } else {
         this.updateItem.emit(result);
@@ -144,7 +149,17 @@ export class GenericCrudTableComponent implements OnInit {
   }
 
   onDeleteItem(item: any) {
-    this.deleteItem.emit(item);
+    this.utilsService
+      .openConfirmModal(
+        'Confirmation',
+        'Cela va supprimer toutes les sièges de cette séance ainsi que les bookings associés à cette séance. Êtes-vous sûr de vouloir supprimer cette entrée ?',
+        ['Confirmer', 'Annuler']
+      )
+      .subscribe((response) => {
+        if (response) {
+          this.deleteItem.emit(item);
+        }
+      });
   }
 
   getColumnName(column: string | { name: string; type: string }): string {
